Rename file organization-nav.vue to OrganizationNav.vue
