+++ components/HeaderNav.vue
@@ -58,7 +58,7 @@
 import { mapState } from 'vuex'
 import MainNav from '@/components/MainNav'
 import BurgerMenuNav from '@/components/BurgerMenuNav'
-import OrganizationNav from '~/components/organization-nav'
+import OrganizationNav from '@/components/OrganizationNav'
 
 export default {
   name: 'HeaderNav',
